+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.7.0
 
+* Adds support for setting the image file format. See `CameraPlatform.setImageFileFormat`.
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
 ## 2.6.0